Pipeline.Error

Description

This package contains the tools needed to throw errors inside a Task.

Errors will be propagated through the network.

Documentation

newtype TaskError

Standard type of an error that is propagated through the network.

Constructors

TaskError ExceptionMessage

newtype ExceptionMessage

A wrapper for an error message

Constructors

ExceptionMessage

Fields

unExceptionMessage :: String

data ExceptT e (m :: Type -> Type) a

A monad transformer that adds exceptions to other monads.

ExceptT constructs a monad parameterized over two things:

e - The exception type.

m - The inner monad.

The return function yields a computation that produces the given value, while >>= sequences two subcomputations, exiting on the first exception.

data SomeException

The SomeException type is the root of the exception type hierarchy. When an exception of type e is thrown, behind the scenes it is encapsulated in a SomeException.

throwE :: forall (m :: Type -> Type) e a. Monad m => e -> ExceptT e m a

Signal an exception value e.

runExceptT (throwE e) = return (Left e)

throwE e >>= m = throwE e

throw :: forall (r :: RuntimeRep) (a :: TYPE r) e. Exception e => e -> a

Throw an exception. Exceptions may be thrown from purely functional code, but may only be caught within the IO monad.

throwIO :: Exception e => e -> IO a

A variant of throw that can only be used within the IO monad.

Although throwIO has a type that is an instance of the type of throw, the two functions are subtly different:

throw e   `seq` x  ===> throw e
throwIO e `seq` x  ===> x

The first example will cause the exception e to be raised, whereas the second one won't. In fact, throwIO will only cause an exception to be raised when it is used within the IO monad. The throwIO variant should be used in preference to throw to raise an exception within the IO monad because it guarantees ordering with respect to other IO operations, whereas throw does not.