Pipeline.Internal.Backend.BasicNetwork

Documentation

data BasicNetwork (inputsStorageType :: [Type -> Type]) (inputsType :: [Type]) (outputsStorageType :: [Type -> Type]) (outputsType :: [Type]) where

Main type for storing information about the process network.

The type arguments align with the Circuit data type.

Constructors

BasicNetwork

Fields

:: { threads :: Map TaskUUID ThreadId

, jobs :: Map JobUUID JobStatus

, inputs :: PipeList inputsStorage inputsType

, outputs :: PipeList outputsStorage outputsType

} -> BasicNetwork inputsStorage inputsType outputsStorage outputsType

data JobStatus

Constructors

Processing

Complete

taskExecuter :: Task iF inputsS inputsT outputS outputT ninputs -> TaskUUID -> PipeList inputsS inputsT -> PipeList outputS outputT -> IO ()

intercept :: ExceptT SomeException IO a -> ExceptT SomeException IO a

writePipes :: JobUUID -> Either TaskError (HList' inputsS inputsT) -> PipeList inputsS inputsT -> IO ()

readPipes :: PipeList outputsS outputsT -> IO (JobUUID, Either TaskError (HList' outputsS outputsT))

initialNetwork :: forall inputsS inputsT. InitialPipes inputsS inputsT => IO (BasicNetwork inputsS inputsT inputsS inputsT)

Creates the initial network used in the accumulating fold

circuitInputs :: (Length bsS ~ Length bsT, ninputs ~ Length bsS, IsNat ninputs, Network n) => N n asS asT bsS bsT csS csT (ninputs :: Nat) -> SNat (Length bsS)

buildBasicNetwork :: InitialPipes a b => Circuit a b c d e -> IO (BasicNetwork a b c d)

beside :: forall asS asT bsS bsT csS csT (nbs :: Nat). Beside (N BasicNetwork asS asT) bsS bsT csS csT nbs -> IO (N BasicNetwork asS asT bsS bsT csS csT nbs)