Pipeline.Internal.Backend.Network

Documentation

class Network n where

Network typeclass

Methods

startNetwork

Arguments

:: InitialPipes inS inT

=> Circuit inS inT outS outT nin
The Circuit used to create the network

-> IO (n inS inT outS outT)
The created network

Starts a network with the given Circuit as specification.

stopNetwork :: n inS inT outS outT -> IO ()

Stops the given network

read

Arguments

:: n inputsS inputsT outputsS outputsT
The network to retrieve inputs from

-> IO (JobUUID, Either TaskError (HList' outputsS outputsT))
The identifier for the output and the output values | Write a set of inputs into the network

This will read from the outputs of the network.

This is a blocking call, therefore if there are no outputs to be read then the program will deadlock.

write

Arguments

:: JobUUID
A unique identifier for the input values

-> HList' inputsS inputsT
The input values

-> n inputsS inputsT outputsS outputsT
The network to input the values in to

-> IO ()

class InitialPipes (inputsS :: [Type -> Type]) (inputsT :: [Type]) where

Used to build a list of pipes from a list of types.

Methods

initialPipes :: IO (PipeList inputsS inputsT)

newtype N n asS asT a b c d e

Constructors

N

Fields

unN :: n asS asT a b -> IO (n asS asT c d)

class (IFunctor5 iF, Network n) => BuildNetworkAlg n iF where

The accumulating fold to build the network.

Methods

buildNetworkAlg :: iF (N n asS asT) bsS bsT csS csT nbs -> IO (N n asS asT bsS bsT csS csT nbs)