Pipeline.Internal.Common.IFunctor

Documentation

class IFunctor iF where

Methods

imap :: (forall a. f a -> g a) -> iF f a -> iF g a

imapM :: Monad m => (forall a. f a -> m (g a)) -> iF f a -> m (iF g a)

class IFunctor2 iF where

Methods

imap2 :: (forall a b. f a b -> g a b) -> iF f a b -> iF g a b

class IFunctor4 iF where

Methods

imap4 :: (forall a b c d. f a b c d -> g a b c d) -> iF f a b c d -> iF g a b c d

class IFunctor6 iF where

Methods

imap6 :: (forall a b c d e f. f' a b c d e f -> g' a b c d e f) -> iF f' a b c d e f -> iF g' a b c d e f

class IFunctor7 iF where

Methods

imap7 :: (forall a b c d e f g. f' a b c d e f g -> g' a b c d e f g) -> iF f' a b c d e f g -> iF g' a b c d e f g

imapM7 :: Monad m => (forall a b c d e f g. f' a b c d e f g -> m (g' a b c d e f g)) -> iF f' a b c d e f g -> m (iF g' a b c d e f g)

newtype Fix f

Constructors

In (f (Fix f))

newtype IFix iF a

Constructors

IIn (iF (IFix iF) a)

newtype IFix2 iF a b

Constructors

IIn2 (iF (IFix2 iF) a b)

newtype IFix3 iF a b c

Constructors

IIn3 (iF (IFix3 iF) a b c)

newtype IFix4 iF a b c d

Constructors

IIn4 (iF (IFix4 iF) a b c d)

newtype IFix5 iF a b c d e

Constructors

IIn5 (iF (IFix5 iF) a b c d e)

newtype IFix6 iF a b c d e f

Constructors

IIn6 (iF (IFix6 iF) a b c d e f)

newtype IFix7 iF a b c d e f g

Constructors

IIn7 (iF (IFix7 iF) a b c d e f g)

icata :: IFunctor iF => (forall a. iF f a -> f a) -> IFix iF a -> f a

icata2 :: IFunctor2 iF => (forall a b. iF f a b -> f a b) -> IFix2 iF a b -> f a b

icata4 :: IFunctor4 iF => (forall a b c d. iF f a b c d -> f a b c d) -> IFix4 iF a b c d -> f a b c d

icata7 :: IFunctor7 iF => (forall a b c d e f g. iF f' a b c d e f g -> f' a b c d e f g) -> IFix7 iF a b c d e f g -> f' a b c d e f g

icataM7 :: (IFunctor7 iF, Monad m) => (forall a b c d e f g. iF f' a b c d e f g -> m (f' a b c d e f g)) -> IFix7 iF a b c d e f g -> m (f' a b c d e f g)

icataM :: (IFunctor iF, Monad m) => (forall a. iF f a -> m (f a)) -> IFix iF a -> m (f a)

cataM :: (Traversable f, Monad m) => (forall a. f a -> m a) -> Fix f -> m a