Pipeline.Internal.Core.DataStore

Documentation

class DataStore f a where

DataStore that can be defined for each datastore needed to be used.

Methods

fetch :: f a -> IO a

Fetch the value stored in the DataStore

save :: f a -> a -> IO ()

Save a value into the DataStore

The first argument depends on the instance. It may be "empty" or it could be a pointer to a storage location.

empty :: TaskUUID -> JobUUID -> IO (f a)

Returns a new empty store

class DataStore' (fs :: [Type -> Type]) (as :: [Type]) where

When tasks require multiple inputs, they also require a joint DataStore. This class provides this ability.

A user does not need to define instances of this class, however it is useful when defining your own tasks.

Methods

fetch' :: HList' fs as -> IO (HList as)

Fetch the value stored in the DataStore'

save' :: HList' fs as -> HList as -> IO ()

Save a value into the DataStore'

The first argument depends on the instance. It may be "empty" or it could be a pointer to a storage location.

empty' :: TaskUUID -> JobUUID -> IO (HList' fs as)

data Var a

Simple in memory variable store, and unmutable.

emptyVar :: IO (Var a)