Pipeline.Task

Description

Contains the tools needed to create, interact with and stop a network.

Creating a Task

A Task, is one of the key components of this system. They will be converted into threads in the process network.

The function a Task executes has a specific type and a task will need to do multiple things.

The first agument will be a HList' containing all the DataStores, to be fetched from. The second argument is the "empty" DataStores or "pointers" to a storage location, where the result should be saved.

A task should:

use fetch' to retrieve all the inputs from the DataStores.

use save' to save the outputs to the DataStores, this should take the second argument given to the Task.

task

Arguments

:: (DataStore' fs as, DataStore g b, Eq (g b), Show (g b), NFData (g b))

=> (UUID -> HList' fs as -> g b -> ExceptT SomeException IO (g b))
The function a Task will execute.

-> g b
The output DataStore

-> Circuit fs as (Apply fs as) '[g] '[b] '[g b] (Length fs)

Constructor for a task

Smart Constructors

multiInputTask

Arguments

:: (DataStore' fs as, DataStore g b, Eq (g b), Show (g b), NFData (g b), NFData b)

=> (HList as -> b)
The function to execute

-> g b
The output DataStore

-> Circuit fs as (Apply fs as) '[g] '[b] '[g b] (Length fs)

This allows a function with multiple inputs to be converted into a Task.

functionTask

Arguments

:: (DataStore f a, DataStore g b, Eq (g b), Show (g b), NFData (g b), NFData b)

=> (a -> b)
The function to execute

-> g b
The output DataStore

-> Circuit '[f] '[a] '[f a] '[g] '[b] '[g b] ('Succ 'Zero)

This allows a single a -> b to be converted into a Task.

Misc

data HList (xs :: [Type]) where

A heterogeneous list used as input/output to a task.

This is commonly used in the multiInputTask function, which automatically fetches the data from DataStore.

Constructors

HCons :: x -> HList xs -> HList (x ': xs)

HNil :: HList '[]

data HList' (fs :: [Type -> Type]) (as :: [Type]) where

A heterogeneous list used as input/output to a network or task.

Constructors

HCons' :: (Eq (f a), Show (f a)) => f a -> HList' fs as -> HList' (f ': fs) (a ': as)

HNil' :: HList' '[] '[]

data ExceptT e (m :: Type -> Type) a

A monad transformer that adds exceptions to other monads.

ExceptT constructs a monad parameterized over two things:

e - The exception type.

m - The inner monad.

The return function yields a computation that produces the given value, while >>= sequences two subcomputations, exiting on the first exception.