Pipeline.Circuit

Description

This package contains all the constructors needed to build a Circuit.

Main Type

type Circuit = IFix7 CircuitF

The core type for a Circuit. It takes 7 different type arguments

Circuit (inputsStorageTypes  :: [Type -> Type]) (inputsTypes  :: [Type]) (inputsApplied  :: [Type])
        (outputsStorageTypes :: [Type -> Type]) (outputsTypes :: [Type]) (outputsApplied :: [Type])
        (nInputs :: Nat)

inputStorageTypes is a type-list of storage types, for example '[VariableStore, CSVStore].

inputTypes is a type-list of the types stored in the storage, for example '[Int, [(String, Float)]].

inputsApplied is a type-list of the storage types applied to the types stored, for example '[VariableStore Int, CSVStore [(String, Float)]].

outputsStorageTypes, inputTypes and outputsApplied mirror the examples above, but for the outputs instead.

nInputs is a type-level Nat that is the length of all input lists.

Constructors

id :: DataStore' '[f] '[a] => Circuit '[f] '[a] '[f a] '[f] '[a] '[f a] N1

Passes an input through without modifying it.

In diagram form it would look like,

|

replicate :: DataStore' '[f] '[a] => Circuit '[f] '[a] '[f a] '[f, f] '[a, a] '[f a, f a] N1

Duplicates an input.

In diagram form it would look like,

/\

(<->) infixr 4

Arguments

:: (DataStore' fs as, DataStore' gs bs, DataStore' hs cs)

=> Circuit fs as (Apply fs as) gs bs (Apply gs bs) nfs
First circuit (a)

-> Circuit gs bs (Apply gs bs) hs cs (Apply hs cs) ngs
Second circuit (b)

-> Circuit fs as (Apply fs as) hs cs (Apply hs cs) nfs

Usually referred to as "then", this operator joins two levels of a circuit together.

A diagram representing a <-> b or "a then b" can be seen below,

| ... |    -- Any number of inputs
   a
| ... |    -- outputs a ~ inputs b
   b
| ... |    -- Any number of outputs

(<>) infixr 5

Arguments

:: (DataStore' fs as, DataStore' gs bs, DataStore' hs cs, DataStore' is ds, nfs ~ Length fs, IsNat nfs, IsNat nhs, Length fs ~ Length as, Length fs ~ Length (Apply fs as), Length gs ~ Length bs, Length gs ~ Length (Apply gs bs), Length hs ~ Length cs, Length hs ~ Length (Apply hs cs), Length is ~ Length ds, Length is ~ Length (Apply is ds), Take (Length as) (Apply fs as :++ Apply hs cs) ~ Apply fs as, Take (Length as) (as :++ cs) ~ as, Take (Length as) (fs :++ hs) ~ fs, Drop (Length as) (Apply fs as :++ Apply hs cs) ~ Apply hs cs, Drop (Length as) (as :++ cs) ~ cs, Drop (Length as) (fs :++ hs) ~ hs, AppendP gs bs (Apply gs bs) is ds (Apply is ds))

=> Circuit fs as (Apply fs as) gs bs (Apply gs bs) nfs
Left circuit

-> Circuit hs cs (Apply hs cs) is ds (Apply is ds) nhs
Right circuit

-> Circuit (fs :++ hs) (as :++ cs) (Apply fs as :++ Apply hs cs) (gs :++ is) (bs :++ ds) (Apply gs bs :++ Apply is ds) (nfs :+ nhs)

Usually referred to as "beside" or "next to", this operator joins two circuits next to each other.

A diagram representing a <> b or "a next to b" can be seen below,

| ... | ... |  -- inputs  a ++ inputs  b
   a     b
| ... | ... |  -- outputs a ++ outputs b

swap :: DataStore' '[f, g] '[a, b] => Circuit '[f, g] '[a, b] '[f a, g b] '[g, f] '[b, a] '[g b, f a] N2

Swaps to input values around.

In diagram form this would look like,

\/
/\

dropL :: DataStore' '[f, g] '[a, b] => Circuit '[f, g] '[a, b] '[f a, g b] '[g] '[b] '[g b] N2

Takes two values as input and drops the left input.

dropR :: DataStore' '[f, g] '[a, b] => Circuit '[f, g] '[a, b] '[f a, g b] '[f] '[a] '[f a] N2

Takes two values as input and drops the right input.

mapC :: (DataStore' '[f] '[[a]], DataStore g [b], Eq (g [b]), Show (g [b]), Eq a, Show a) => Circuit '[VariableStore] '[a] '[VariableStore a] '[VariableStore] '[b] '[VariableStore b] N1 -> g [b] -> Circuit '[f] '[[a]] '[f [a]] '[g] '[[b]] '[g [b]] N1

Maps a circuit on the inputs