Pipeline.Internal.Core.CircuitAST

Documentation

type Circuit = IFix7 CircuitF

The core type for a Circuit. It takes 7 different type arguments

Circuit (inputsStorageTypes  :: [Type -> Type]) (inputsTypes  :: [Type]) (inputsApplied  :: [Type])
        (outputsStorageTypes :: [Type -> Type]) (outputsTypes :: [Type]) (outputsApplied :: [Type])
        (nInputs :: Nat)

inputStorageTypes is a type-list of storage types, for example '[VariableStore, CSVStore].

inputTypes is a type-list of the types stored in the storage, for example '[Int, [(String, Float)]].

inputsApplied is a type-list of the storage types applied to the types stored, for example '[VariableStore Int, CSVStore [(String, Float)]].

outputsStorageTypes, inputTypes and outputsApplied mirror the examples above, but for the outputs instead.

nInputs is a type-level Nat that is the length of all input lists.

type CircuitF = Id :+: (Replicate :+: (Then :+: (Beside :+: (Swap :+: (DropL :+: (DropR :+: (Task :+: Map)))))))

Constructors

data Id (iF :: [Type -> Type] -> [Type] -> [Type] -> [Type -> Type] -> [Type] -> [Type] -> Nat -> Type) (inputsS :: [Type -> Type]) (inputsT :: [Type]) (inputsA :: [Type]) (outputsS :: [Type -> Type]) (outputsT :: [Type]) (outputsA :: [Type]) (ninputs :: Nat) where

Constructors

Id :: DataStore' '[inputS] '[inputT] => Id iF '[inputS] '[inputT] '[inputS inputT] '[inputS] '[inputT] '[inputS inputT] N1

data Replicate (iF :: [Type -> Type] -> [Type] -> [Type] -> [Type -> Type] -> [Type] -> [Type] -> Nat -> Type) (inputsS :: [Type -> Type]) (inputsT :: [Type]) (inputsA :: [Type]) (outputsS :: [Type -> Type]) (outputsT :: [Type]) (outputsA :: [Type]) (ninputs :: Nat) where

Constructors

Replicate :: DataStore' '[f] '[a] => Replicate iF '[f] '[a] '[f a] '[f, f] '[a, a] '[f a, f a] N1

data Then (iF :: [Type -> Type] -> [Type] -> [Type] -> [Type -> Type] -> [Type] -> [Type] -> Nat -> Type) (inputsS :: [Type -> Type]) (inputsT :: [Type]) (inputsA :: [Type]) (outputsS :: [Type -> Type]) (outputsT :: [Type]) (outputsA :: [Type]) (ninputs :: Nat) where

Constructors

Then :: (DataStore' fs as, DataStore' gs bs, DataStore' hs cs) => iF fs as (Apply fs as) gs bs (Apply gs bs) nfs -> iF gs bs (Apply gs bs) hs cs (Apply hs cs) ngs -> Then iF fs as (Apply fs as) hs cs (Apply hs cs) nfs

data Beside (iF :: [Type -> Type] -> [Type] -> [Type] -> [Type -> Type] -> [Type] -> [Type] -> Nat -> Type) (inputsS :: [Type -> Type]) (inputsT :: [Type]) (inputsA :: [Type]) (outputsS :: [Type -> Type]) (outputsT :: [Type]) (outputsA :: [Type]) (ninputs :: Nat) where

Constructors

Beside :: (DataStore' fs as, DataStore' gs bs, DataStore' hs cs, DataStore' is ds, IsNat nfs, IsNat nhs, nfs ~ Length fs, Length fs ~ Length as, Length fs ~ Length (Apply fs as), Length gs ~ Length bs, Length gs ~ Length (Apply gs bs), Length hs ~ Length cs, Length hs ~ Length (Apply hs cs), Length is ~ Length ds, Length is ~ Length (Apply is ds), Take (Length as) (Apply fs as :++ Apply hs cs) ~ Apply fs as, Take (Length as) (as :++ cs) ~ as, Take (Length as) (fs :++ hs) ~ fs, Drop (Length as) (Apply fs as :++ Apply hs cs) ~ Apply hs cs, Drop (Length as) (as :++ cs) ~ cs, Drop (Length as) (fs :++ hs) ~ hs, AppendP gs bs (Apply gs bs) is ds (Apply is ds)) => iF fs as (Apply fs as) gs bs (Apply gs bs) nfs -> iF hs cs (Apply hs cs) is ds (Apply is ds) nhs -> Beside iF (fs :++ hs) (as :++ cs) (Apply fs as :++ Apply hs cs) (gs :++ is) (bs :++ ds) (Apply gs bs :++ Apply is ds) (nfs :+ nhs)

data Swap (iF :: [Type -> Type] -> [Type] -> [Type] -> [Type -> Type] -> [Type] -> [Type] -> Nat -> Type) (inputsS :: [Type -> Type]) (inputsT :: [Type]) (inputsA :: [Type]) (outputsS :: [Type -> Type]) (outputsT :: [Type]) (outputsA :: [Type]) (ninputs :: Nat) where

Constructors

Swap :: DataStore' '[f, g] '[a, b] => Swap iF '[f, g] '[a, b] '[f a, g b] '[g, f] '[b, a] '[g b, f a] N2

data DropL (iF :: [Type -> Type] -> [Type] -> [Type] -> [Type -> Type] -> [Type] -> [Type] -> Nat -> Type) (inputsS :: [Type -> Type]) (inputsT :: [Type]) (inputsA :: [Type]) (outputsS :: [Type -> Type]) (outputsT :: [Type]) (outputsA :: [Type]) (ninputs :: Nat) where

Constructors

DropL :: DataStore' '[f, g] '[a, b] => DropL iF '[f, g] '[a, b] '[f a, g b] '[g] '[b] '[g b] N2

data DropR (iF :: [Type -> Type] -> [Type] -> [Type] -> [Type -> Type] -> [Type] -> [Type] -> Nat -> Type) (inputsS :: [Type -> Type]) (inputsT :: [Type]) (inputsA :: [Type]) (outputsS :: [Type -> Type]) (outputsT :: [Type]) (outputsA :: [Type]) (ninputs :: Nat) where

Constructors

DropR :: DataStore' '[f, g] '[a, b] => DropR iF '[f, g] '[a, b] '[f a, g b] '[f] '[a] '[f a] N2

data Task (iF :: [Type -> Type] -> [Type] -> [Type] -> [Type -> Type] -> [Type] -> [Type] -> Nat -> Type) (inputsS :: [Type -> Type]) (inputsT :: [Type]) (inputsA :: [Type]) (outputsS :: [Type -> Type]) (outputsT :: [Type]) (outputsA :: [Type]) (ninputs :: Nat) where

Constructors

Task :: ((Length outputsS := 'Succ 'Zero) ~ 'True, outputsS ~ '[g'], outputsT ~ '[b'], outputsA ~ '[g' b'], DataStore' inputsS inputsT, DataStore g' b', Eq (g' b'), Show (g' b'), NFData (g' b')) => (UUID -> HList' inputsS inputsT -> g' b' -> ExceptT SomeException IO (g' b')) -> g' b' -> Task iF inputsS inputsT (Apply inputsS inputsT) outputsS outputsT outputsA (Length inputsS)

data Map (iF :: [Type -> Type] -> [Type] -> [Type] -> [Type -> Type] -> [Type] -> [Type] -> Nat -> Type) (inputsS :: [Type -> Type]) (inputsT :: [Type]) (inputsA :: [Type]) (outputsS :: [Type -> Type]) (outputsT :: [Type]) (outputsA :: [Type]) (ninputs :: Nat) where

Constructors

Map :: (DataStore' '[f] '[[a]], DataStore g [b], Eq (g [b]), Show (g [b]), Eq a, Show a) => Circuit '[VariableStore] '[a] '[VariableStore a] '[VariableStore] '[b] '[VariableStore b] N1 -> g [b] -> Map iF '[f] '[[a]] '[f [a]] '[g] '[[b]] '[g [b]] N1