Pipeline.Network

Description

Contains the tools needed to create, interact with and stop a network.

Network

class Network n where

Network typeclass

Methods

startNetwork

Arguments

:: InitialPipes inS inT inA

=> Circuit inS inT inA outS outT outA nin
The Circuit used to create the network

-> IO (n inS inT inA outS outT outA)
The created network

Starts a network with the given Circuit as specification.

stopNetwork :: n inS inT inA outS outT outA -> IO ()

Stops the given network

read

Arguments

:: n inputsS inputsT inputsA outputsS outputsT outputsA
The network to retrieve inputs from

-> IO (UUID, Either TaskError (HList' outputsS outputsT))
The identifier for the output and the output values | Write a set of inputs into the network

This will read from the outputs of the network.

This is a blocking call, therefore if there are no outputs to be read then the program will deadlock.

write

Arguments

:: UUID
A unique identifier for the input values

-> HList' inputsS inputsT
The input values

-> n inputsS inputsT inputsA outputsS outputsT outputsA
The network to input the values in to

-> IO ()

Basic Network

data BasicNetwork (inputsStorageType :: [Type -> Type]) (inputsType :: [Type]) (inputsAp :: [Type]) (outputsStorageType :: [Type -> Type]) (outputsType :: [Type]) (outputsAp :: [Type])

Main type for storing information about the process network.

The type arguments align with the Circuit data type.

Network IO

type UUID = String

A synonym for a Unique Id for a task

Input

input

Arguments

:: Network n

=> HList' inputsS inputsT
Inputs to the network

-> n inputsS inputsT inputsA outputsS outputsT outputsA
Network to insert the values in

-> IO UUID
Randomly generated identifier

Input values into a network. This will return a randomly generated identifier for the inputs.

input_ :: HList' inputsS inputsT -> BasicNetwork inputsS inputsT inputsA outputsS outputsT outputsA -> IO ()

A variant of input, however it will not return the randomly generated identifier.

Output

output_ :: Network n => n inputsS inputsT inputsA outputsS outputsT outputsA -> IO (Either TaskError (HList' outputsS outputsT))

A variant of read, that does not return the unique identifier.

This is a blocking call, therefore if there are no outputs to be read then the program will deadlock.

data HList' (fs :: [Type -> Type]) (as :: [Type]) where

A heterogeneous list used as input/output to a network or task.

Constructors

HCons' :: (Eq (f a), Show (f a)) => f a -> HList' fs as -> HList' (f ': fs) (a ': as)

HNil' :: HList' '[] '[]

Misc

class InitialPipes (inputsS :: [Type -> Type]) (inputsT :: [Type]) (inputsA :: [Type])

Used to build a list of pipes from a list of types.

Minimal complete definition

initialPipes